// src/app/schemas/map-name-schemas.ts
export enum SurvivingMarsMapName {
  BlankBig_01 = 'BlankBig_01',
  BlankBig_02 = 'BlankBig_02',
  BlankBig_03 = 'BlankBig_03',
  BlankBig_04 = 'BlankBig_04',
  BlankBigCanyonCMix_01 = 'BlankBigCanyonCMix_01',
  BlankBigCanyonCMix_02 = 'BlankBigCanyonCMix_02',
  BlankBigCanyonCMix_03 = 'BlankBigCanyonCMix_03',
  BlankBigCanyonCMix_04 = 'BlankBigCanyonCMix_04',
  BlankBigCanyonCMix_05 = 'BlankBigCanyonCMix_05',
  BlankBigCanyonCMix_06 = 'BlankBigCanyonCMix_06',
  BlankBigCanyonCMix_07 = 'BlankBigCanyonCMix_07',
  BlankBigCanyonCMix_08 = 'BlankBigCanyonCMix_08',
  BlankBigCanyonCMix_09 = 'BlankBigCanyonCMix_09',
  BlankBigCanyonCMix_10 = 'BlankBigCanyonCMix_10',
  BlankBigCliffsCMix_01 = 'BlankBigCliffsCMix_01',
  BlankBigCliffsCMix_02 = 'BlankBigCliffsCMix_02',
  BlankBigCrater_01 = 'BlankBigCrater_01',
  BlankBigCraterCMix_01 = 'BlankBigCraterCMix_01',
  BlankBigCraterCMix_02 = 'BlankBigCraterCMix_02',
  BlankBigHeartCMix_03 = 'BlankBigHeartCMix_03',
  BlankBigTerraceCMix_01 = 'BlankBigTerraceCMix_01',
  BlankBigTerraceCMix_02 = 'BlankBigTerraceCMix_02',
  BlankBigTerraceCMix_03 = 'BlankBigTerraceCMix_03',
  BlankBigTerraceCMix_04 = 'BlankBigTerraceCMix_04',
  BlankBigTerraceCMix_05 = 'BlankBigTerraceCMix_05',
  BlankBigTerraceCMix_06 = 'BlankBigTerraceCMix_06',
  BlankBigTerraceCMix_07 = 'BlankBigTerraceCMix_07',
  BlankBigTerraceCMix_08 = 'BlankBigTerraceCMix_08',
  BlankBigTerraceCMix_09 = 'BlankBigTerraceCMix_09',
  BlankBigTerraceCMix_10 = 'BlankBigTerraceCMix_10',
  BlankBigTerraceCMix_11 = 'BlankBigTerraceCMix_11',
  BlankBigTerraceCMix_12 = 'BlankBigTerraceCMix_12',
  BlankBigTerraceCMix_13 = 'BlankBigTerraceCMix_13',
  BlankBigTerraceCMix_14 = 'BlankBigTerraceCMix_14',
  BlankBigTerraceCMix_15 = 'BlankBigTerraceCMix_15',
  BlankBigTerraceCMix_16 = 'BlankBigTerraceCMix_16',
  BlankBigTerraceCMix_17 = 'BlankBigTerraceCMix_17',
  BlankBigTerraceCMix_18 = 'BlankBigTerraceCMix_18',
  BlankBigTerraceCMix_19 = 'BlankBigTerraceCMix_19',
  BlankBigTerraceCMix_20 = 'BlankBigTerraceCMix_20',
  BlankTerraceBig_05 = 'BlankTerraceBig_05',
  /* ------ BELOW AND BEYOND DLC MAPS ------ */
  BlankUnderground_01 = 'BlankUnderground_01',
  BlankUnderground_02 = 'BlankUnderground_02',
  BlankUnderground_03 = 'BlankUnderground_03',
  BlankUnderground_04 = 'BlankUnderground_04',
}

export type SurvivingMarsMapId = keyof typeof SurvivingMarsMapName;